+++ cvat-ui/src/components/email-confirmation-pages/email-confirmed.tsx
