+++ cvat-ui/src/components/email-confirmation-pages/styles.scss
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 #email-confirmation-page-container,
-#email-verification-is-required-page-container {
+#email-verification-sent-page-container {
     height: 100%;
     text-align: center;
 }
