+++ cvat-ui/src/components/login-page/login-page.tsx
@@ -6,12 +6,18 @@
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { Link, withRouter } from 'react-router-dom';
+import Button from 'antd/lib/button';
 import Title from 'antd/lib/typography/Title';
 import Text from 'antd/lib/typography/Text';
 import { Row, Col } from 'antd/lib/grid';
 import Layout from 'antd/lib/layout';
+import Space from 'antd/lib/space';
+import { GithubOutlined, GooglePlusOutlined } from '@ant-design/icons';
 
 import LoginForm, { LoginData } from './login-form';
+import { getCore } from '../../cvat-core-wrapper';
+
+const cvat = getCore();
 
 interface LoginPageComponentProps {
     fetching: boolean;
@@ -20,6 +26,7 @@ interface LoginPageComponentProps {
 }
 
 function LoginPageComponent(props: LoginPageComponentProps & RouteComponentProps): JSX.Element {
+    const { backendAPI } = cvat.config;
     const sizes = {
         style: {
             width: 400,
@@ -43,6 +50,29 @@ function LoginPageComponent(props: LoginPageComponentProps & RouteComponentProps
                                 onLogin(loginData.credential, loginData.password);
                             }}
                         />
+                        <Row justify='center' align='top'>
+                            <Col>
+                                or
+                            </Col>
+                        </Row>
+                        <Row justify='space-between' align='middle'>
+                            <Col span={11}>
+                                <Button href={`${backendAPI}/auth/google/login`}>
+                                    <Space>
+                                        <GooglePlusOutlined />
+                                        Continue with Google
+                                    </Space>
+                                </Button>
+                            </Col>
+                            <Col span={11} offset={1}>
+                                <Button href={`${backendAPI}/auth/github/login`}>
+                                    <Space>
+                                        <GithubOutlined />
+                                        Continue with Github
+                                    </Space>
+                                </Button>
+                            </Col>
+                        </Row>
                         <Row justify='start' align='top'>
                             <Col>
                                 <Text strong>
