+++ cvat/settings/base.py
@@ -265,7 +265,7 @@ ACCOUNT_AUTHENTICATION_METHOD = 'username_email'
 # set UI url to redirect after a successful e-mail confirmation
 #changed from '/auth/login' to '/auth/email-confirmation' for email confirmation message
 ACCOUNT_EMAIL_CONFIRMATION_ANONYMOUS_REDIRECT_URL = '/auth/email-confirmation'
-ACCOUNT_EMAIL_VERIFICATION_IS_REQUIRED_REDIRECT_URL = '/auth/account-email-verification-sent'
+ACCOUNT_EMAIL_VERIFICATION_SENT_REDIRECT_URL = '/auth/email-verification-sent'
 
 OLD_PASSWORD_FIELD_ENABLED = True
