+++ cvat/settings/development.py
@@ -39,7 +39,7 @@ if UI_PORT and UI_PORT != '80':
     UI_URL += ':{}'.format(UI_PORT)
 # set UI url to redirect to after successful e-mail confirmation
 ACCOUNT_EMAIL_CONFIRMATION_ANONYMOUS_REDIRECT_URL = '{}/auth/email-confirmation'.format(UI_URL)
-ACCOUNT_EMAIL_VERIFICATION_IS_REQUIRED_REDIRECT_URL = '{}/auth/account-email-verification-sent'.format(UI_URL)
+ACCOUNT_EMAIL_VERIFICATION_SENT_REDIRECT_URL = '{}/auth/email-verification-sent'.format(UI_URL)
 
 CORS_ORIGIN_WHITELIST = [UI_URL]
 CORS_REPLACE_HTTPS_REFERER = True
